update: reorganize packages
