--- @core/authentication-service/src/main/java/org/flowio/authenticationservice/Entities/Role.java
@@ -1,6 +1,0 @@
-package org.flowio.authenticationservice.Entities;
-
-public enum Role {
-    USER,
-    ADMIN
-}
