+++ @core/authentication-service/src/main/java/org/flowio/authenticationservice/controller/DemoController.java
@@ -1,15 +1,15 @@
-package org.flowio.authenticationservice.demo;
+package org.flowio.authenticationservice.controller;
 
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/api/v1/demo-controller")
+@RequestMapping("/api/v1/ping")
 public class DemoController {
 
     @GetMapping
-    public ResponseEntity<String> sayHello() {
+    public ResponseEntity<String> pingPong() {
         return ResponseEntity.ok("Hello from secured endpoint");
     }
