+++ @core/authentication-service/src/main/java/org/flowio/authenticationservice/dto/AuthenticationResponse.java
@@ -1,4 +1,4 @@
-package org.flowio.authenticationservice.auth;
+package org.flowio.authenticationservice.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
