+++ @core/authentication-service/src/main/java/org/flowio/authenticationservice/model/Role.java
@@ -1,0 +1,6 @@
+package org.flowio.authenticationservice.model;
+
+public enum Role {
+    USER,
+    ADMIN
+}
