+++ @core/authentication-service/src/main/java/org/flowio/authenticationservice/repository/LoginUserRepository.java
@@ -1,5 +1,6 @@
-package org.flowio.authenticationservice.Entities;
+package org.flowio.authenticationservice.repository;
 
+import org.flowio.authenticationservice.model.LoginUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
